:octocat: WIP
It appears that octocrab has the limitation of not supporting
the actual download of files through the github API, or
maybe its a limitation of the github API its self, either way,
it looks like I need another library. Using octocrab github api
calls to search and build the data that shran needs, and a curl
like library to perform the actual downloads.

[issue: 12]